requirements.txt
scipy
sqlalchemy
python-dotenv
polars>=1.0.0
numpy
matplotlib
pyarrow
